Fixes to remove warnings on WIN64 compilation
git-svn-id: https://www.h3d.org:8090/HAPI/trunk@465 f896c5a5-350d-0410-967e-bc877d8f5295

+++ OpenHapticsRenderer/src/OpenHapticsRenderer.cpp
@@ -556,7 +556,7 @@ HLuint OpenHapticsRenderer::getHLShapeId( HAPIHapticShape *hs,
 
     addHLEventCallbacks( hl_shape_id, cb_data );
 
-    id_cb_map[ key ] = callback_data.size() - 1;
+    id_cb_map[ key ] = (unsigned int) callback_data.size() - 1;
   } else {
     unsigned int index = id_cb_map[ key ];
     callback_data[ index ]->shape.reset( hs );
